Incorporate Servers into Server Groups topology tab
Incorporate the servers information (details about each server) onto
the exising ServerGroups tab, and rename the resulting tab (and file)
to be just Servers.

Update hyperlinks from other topology pages.
