+++ src/utils/RouteConfig.js
@@ -31,7 +31,7 @@ import ControlPlanes from '../pages/topology/ControlPlanes.js';
 import Regions from '../pages/topology/Regions.js';
 import Services from '../pages/topology/Services.js';
 import Network from '../pages/topology/Networks.js';
-import ServerGroups from '../pages/topology/ServerGroups.js';
+import Servers from '../pages/topology/Servers.js';
 import ServerRoles from '../pages/topology/ServerRoles.js';
 import { isProduction } from './ConfigHelper.js';
 
@@ -70,7 +70,7 @@ export const routes = [
       { name: translate('regions'), slug: '/topology/regions', component: Regions },
       { name: translate('services'), slug: '/topology/services', component: Services },
       { name: translate('networks'), slug: '/topology/networks', component: Network },
-      { name: translate('server_groups'), slug: '/topology/server-groups', component: ServerGroups },
+      { name: translate('servers'), slug: '/topology/servers', component: Servers },
       { name: translate('roles'), slug: '/topology/roles', component: ServerRoles },
     ]
   },
